add filter_sources_and_artifacts
